+++ server/api/v0/updateaccountid/index.ts
@@ -57,6 +57,7 @@ export const POST = apiHandler("POST", async (req, res) => {
 		mfaRecoverySeed: storedUser.mfaRecoverySeed ?? null,
 		passwordHash: storedUser.passwordHash,
 		passwordSalt: storedUser.passwordSalt,
+		pubnubCipherKey: storedUser.pubnubCipherKey,
 		requiredAddtlAuth: storedUser.requiredAddtlAuth ?? [],
 		totpSeed: storedUser.totpSeed ?? null,
 		uid: storedUser.uid,