+++ server/auth/__mocks__/jwt.ts
@@ -1,8 +1,10 @@
 import type {
 	blacklistHasJwt as _blacklistHasJwt,
 	addJwtToBlacklist as _addJwtToBlacklist,
-	newAccessToken as _newAccessToken,
-	jwtTokenFromRequest as _jwtTokenFromRequest,
+	newAccessTokens as _newAccessTokens,
+	setSession as _setSession,
+	killSession as _killSession,
+	jwtFromRequest as _jwtFromRequest,
 	verifyJwt as _verifyJwt,
 } from "../jwt";
 import { jest } from "@jest/globals";
@@ -13,9 +15,15 @@ export const blacklistHasJwt = jest.fn<typeof _blacklistHasJwt>().mockResolvedVa
 
 export const addJwtToBlacklist = jest.fn<typeof _addJwtToBlacklist>().mockResolvedValue(undefined);
 
-export const newAccessToken = jest.fn<typeof _newAccessToken>().mockResolvedValue("deft");
+export const newAccessTokens = jest
+	.fn<typeof _newAccessTokens>()
+	.mockResolvedValue({ access_token: "deft", pubnub_token: "move" });
 
-export const jwtTokenFromRequest = jest.fn<typeof _jwtTokenFromRequest>().mockReturnValue(null);
+export const setSession = jest.fn<typeof _setSession>();
+
+export const killSession = jest.fn<typeof _killSession>();
+
+export const jwtFromRequest = jest.fn<typeof _jwtFromRequest>().mockReturnValue(null);
 
 export const verifyJwt = jest
 	.fn<typeof _verifyJwt>()