+++ server/environment.ts
@@ -4,7 +4,10 @@ export type EnvKey =
 	| "AUTH_SECRET" //
 	| "HOST"
 	| "MAX_BYTES"
-	| "MAX_USERS";
+	| "MAX_USERS"
+	| "PUBNUB_PUBLISH_KEY"
+	| "PUBNUB_SUBSCRIBE_KEY"
+	| "PUBNUB_SECRET_KEY";
 
 /**
  * Retrieves an environment value for the given key, or `undefined` if none was set.