+++ src/@types/env.d.ts
@@ -3,6 +3,7 @@ interface ImportMetaEnv extends Readonly<Record<string, string>> {
 	readonly VITE_ACCOUNTABLE_SERVER_URL: string | undefined;
 	readonly VITE_ENABLE_SIGNUP: string | undefined;
 	readonly VITE_ENABLE_LOGIN: string | undefined;
+	readonly VITE_PUBNUB_SUBSCRIBE_KEY: string | undefined;
 }
 
 interface ImportMeta {