+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- All API endpoints now support being run as Vercel serverless functions.
+- When running under Vercel, the server cannot handle WebSocket requests directly. Clients must use the `PubNub` client to subscribe to event channels.
+- Server administrators who wish to host on Vercel should set up a PubNub account, as described in the server's [README](/server/README.md#prerequesites).
+
 ## [0.14.0] - 2022-10-02
 ### Changed
 - The server's `DB` environment variable is now ignored. All persistent data now lives on the database identified by the value of the server's `DATABASE_URL` environment variable. [See the README](/server/README.md) for details.